+++ script/provider/provider.lua
@@ -18,6 +18,8 @@ local lang      = require 'language'
 local plugin    = require 'plugin'
 local progress  = require 'progress'
 local tm        = require 'text-merger'
+local vm        = require 'vm'
+local nonil     = require 'without-check-nil'
 
 local function updateConfig()
     local diagnostics = require 'provider.diagnostic'
@@ -40,6 +42,10 @@ local function updateConfig()
                 scopeUri = workspace.uri,
                 section = 'editor.semanticHighlighting.enabled',
             },
+            {
+                scopeUri = workspace.uri,
+                section = 'editor.acceptSuggestionOnEnter',
+            },
         },
     })
     if not configs or not configs[1] then
@@ -49,9 +55,10 @@ local function updateConfig()
 
     local updated = configs[1]
     local other   = {
-        associations = configs[2],
-        exclude      = configs[3],
-        semantic     = configs[4],
+        associations            = configs[2],
+        exclude                 = configs[3],
+        semantic                = configs[4],
+        acceptSuggestionOnEnter = configs[5],
     }
 
     local oldConfig = util.deepCopy(config.config)
@@ -115,29 +122,38 @@ proto.on('initialized', function (params)
     updateConfig()
     local registrations = {}
 
-    -- 监视文件变化
-    registrations[#registrations+1] = {
-        id = 'workspace/didChangeWatchedFiles',
-        method = 'workspace/didChangeWatchedFiles',
-        registerOptions = {
-            watchers = {
-                {
-                    globPattern = '**/',
-                    kind = 1 | 2 | 4,
-                }
+    nonil.enable()
+    if client.info.capabilities.workspace.didChangeWatchedFiles.dynamicRegistration then
+        -- 监视文件变化
+        registrations[#registrations+1] = {
+            id = 'workspace/didChangeWatchedFiles',
+            method = 'workspace/didChangeWatchedFiles',
+            registerOptions = {
+                watchers = {
+                    {
+                        globPattern = '**/',
+                        kind = 1 | 2 | 4,
+                    }
+                },
             },
-        },
-    }
+        }
+    end
 
-    -- 监视配置变化
-    registrations[#registrations+1] = {
-        id = 'workspace/didChangeConfiguration',
-        method = 'workspace/didChangeConfiguration',
-    }
+    if client.info.capabilities.workspace.didChangeConfiguration.dynamicRegistration then
+        -- 监视配置变化
+        registrations[#registrations+1] = {
+            id = 'workspace/didChangeConfiguration',
+            method = 'workspace/didChangeConfiguration',
+        }
+    end
 
-    proto.awaitRequest('client/registerCapability', {
-        registrations = registrations
-    })
+    nonil.disable()
+
+    if #registrations ~= 0 then
+        proto.awaitRequest('client/registerCapability', {
+            registrations = registrations
+        })
+    end
     workspace.reload()
     return true
 end)
@@ -288,6 +304,7 @@ proto.on('textDocument/hover', function (params)
     if not files.exists(uri) then
         return nil
     end
+    vm.flushCache()
     local offset = files.offsetOfWord(uri, params.position)
     local hover = core.byUri(uri, offset)
     if not hover then
@@ -314,6 +331,7 @@ proto.on('textDocument/definition', function (params)
     if not files.exists(uri) then
         return nil
     end
+    vm.flushCache()
     local offset = files.offsetOfWord(uri, params.position)
     local result = core(uri, offset)
     if not result then
@@ -343,6 +361,7 @@ proto.on('textDocument/references', function (params)
     if not files.exists(uri) then
         return nil
     end
+    vm.flushCache()
     local offset = files.offsetOfWord(uri, params.position)
     local result = core(uri, offset)
     if not result then
@@ -364,6 +383,7 @@ proto.on('textDocument/documentHighlight', function (params)
     if not files.exists(uri) then
         return nil
     end
+    vm.flushCache()
     local offset = files.offsetOfWord(uri, params.position)
     local result = core(uri, offset)
     if not result then
@@ -387,6 +407,7 @@ proto.on('textDocument/rename', function (params)
     if not files.exists(uri) then
         return nil
     end
+    vm.flushCache()
     local offset = files.offsetOfWord(uri, params.position)
     local result = core.rename(uri, offset, params.newName)
     if not result then
@@ -412,6 +433,7 @@ proto.on('textDocument/prepareRename', function (params)
     if not files.exists(uri) then
         return nil
     end
+    vm.flushCache()
     local offset = files.offsetOfWord(uri, params.position)
     local result = core.prepareRename(uri, offset)
     if not result then
@@ -435,6 +457,13 @@ proto.on('textDocument/completion', function (params)
     local uri  = params.textDocument.uri
     if not files.exists(uri) then
         return nil
+    end
+    if config.other.acceptSuggestionOnEnter ~= 'off' then
+        if params.context.triggerCharacter == '\n'
+        or params.context.triggerCharacter == '{'
+        or params.context.triggerCharacter == ',' then
+            return
+        end
     end
     await.setPriority(1000)
     local clock  = os.clock()