+++ changelog.md
@@ -1,5 +1,10 @@
 # changelog
 
+## 1.21.0
+* `CHG` `LuaDoc`: supports `---@param self TYPE`
+* `CHG` completion: does not show suggests after `\n`, `{` and `,`, unless your setting `editor.acceptSuggestionOnEnter` is `off`
+* `FIX` [#482](https://github.com/sumneko/lua-language-server/issues/482)
+
 ## 1.20.1
 `2021-3-27`
 * `FIX` telemetry window blocks initializing